Show if a version is EOL or not in the release details modal
