+++ src/stores/releases/utils.ts
@@ -1,0 +1,30 @@
+import { compareDates, getRelativeDateFromNow } from 'lib/helpers';
+
+export function getReleaseEOLStatus(
+  eolDate: string
+): { message: string; isEol: boolean } {
+  const result = {
+    message: '',
+    isEol: false,
+  };
+
+  if (!eolDate) return result;
+
+  const now = new Date().toISOString();
+  const relativeDate = getRelativeDateFromNow(eolDate);
+  switch (compareDates(now, eolDate)) {
+    case -1:
+      result.message = `This version will reach its end of life ${relativeDate}.`;
+      break;
+    case 0:
+      result.message = 'This version reached its end of life today.';
+      result.isEol = true;
+      break;
+    case 1:
+      result.message = `This version reached its end of life ${relativeDate}.`;
+      result.isEol = true;
+      break;
+  }
+
+  return result;
+}
